refactor(mountains): introduce src/mountains/ facade + index hub
Move src/mountains.ts wholesale into src/mountains/index.ts and replace
src/mountains.ts with a thin root facade (export * from './mountains/index.js'),
mirroring the src/snowman.ts -> src/snowman/index.ts split (Stage R3, #34).

No behavior change: every importer keeps resolving the sibling ./mountains.js
specifier, which now re-exports the Mountains object, the named samplers/helpers,
and the TerrainVec2/RockPosition types. The implementation hub will be carved into
focused submodules (noise / terrain / snow-surface / terrain-mesh / rocks) in the
following commits; terrain math stays byte-identical at every step.
